backend/debt_service/serializers.py
from rest_framework import serializers
from .models import Debt, UploadedFile

# Transform model instances to JSON.
class DebtSerializer(serializers.ModelSerializer):
    class Meta:
        model = Debt
        fields = '__all__'

class UploadedFileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UploadedFile
        fields = '__all__'
