backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
celery==5.1.2
redis==3.5.3
pyspark==3.1.2
pika==1.2.0
#pyboleto==0.3
#psycopg2==2.8.6
pytest==6.2.4
gunicorn==20.1.0
